+++ hal/src/sercom/uart/flags.rs
@@ -77,6 +77,7 @@ bitflags! {
 
 /// Errors available for UART transactions
 #[derive(Debug, Clone, Copy)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Error {
     /// Detected a parity error
     ParityError,