+++ hal/src/thumbv7em/dsu.rs
@@ -15,6 +15,7 @@ pub struct Dsu {
 
 /// Errors from hardware
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum PeripheralError {
     /// Usually misaligned address of length
     BusError,
@@ -22,6 +23,7 @@ pub enum PeripheralError {
 
 /// Error from within the DSU
 #[derive(Debug)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub enum Error {
     /// Address or length was not word aligned
     AlignmentError,